shared/server/requirements.txt
# Install Django
Django==1.8.6

# Required for Django timezones. This is from the django docs: https://docs.djangoproject.com/en/1.8/topics/i18n/timezones/
pytz

# Postgres Database Adapter (requires 'python-dev' and 'libpq-dev' packages in Ubuntu)
psycopg2

# Database config loader
dj-database-url

# Install rest-api framework
djangorestframework

# Additional packages for rest-api framework
markdown
django-filter
django-crispy-forms
django-guardian